refactor(logic): move frontend code to shared
