+++ src/shared/util/__tests__/logic.spec.ts
@@ -1,12 +1,15 @@
 import { ObjectId } from 'bson-ext'
 
 import {
+  getApplicableIfFields,
+  getApplicableIfStates,
   getLogicUnitPreventingSubmit,
   getVisibleFieldIds,
 } from 'src/shared/util/logic'
 import {
   BasicField,
   FieldResponse,
+  IField,
   IFieldSchema,
   IFormDocument,
   IPreventSubmitLogicSchema,
@@ -999,6 +1002,79 @@ describe('Logic validation', () => {
           form,
         ),
       ).toBeUndefined()
+    })
+  })
+})
+
+describe('Logic util', () => {
+  const VALID_IF_CONDITION_FIELDS = [
+    BasicField.Dropdown,
+    BasicField.Number,
+    BasicField.Decimal,
+    BasicField.Rating,
+    BasicField.YesNo,
+    BasicField.Radio,
+  ]
+
+  const INVALID_IF_CONDITION_FIELDS = Object.values(BasicField).filter(
+    (fieldType) => !VALID_IF_CONDITION_FIELDS.includes(fieldType),
+  )
+
+  describe('getApplicableIfFields', () => {
+    it('should not filter fields suitable as an if-conditional', () => {
+      const validIfFields: IField[] = VALID_IF_CONDITION_FIELDS.map(
+        (fieldType) => ({ fieldType } as unknown as IField),
+      )
+      const fields = getApplicableIfFields(validIfFields)
+      validIfFields.forEach((v, i) => expect(v).toStrictEqual(fields[i]))
+    })
+    it('should filter fields not suitable as an if-conditional', () => {
+      const invalidIfFields: IField[] = INVALID_IF_CONDITION_FIELDS.map(
+        (x) => x as unknown as IField,
+      )
+      const fields = getApplicableIfFields(invalidIfFields)
+      expect(fields).toStrictEqual([])
+    })
+  })
+
+  describe('getApplicableIfStates', () => {
+    it('should return valid logic states for categorical field types', () => {
+      const categoricalFields = [BasicField.Dropdown, BasicField.Radio]
+      categoricalFields.forEach((fieldType) => {
+        const states = getApplicableIfStates(fieldType)
+        expect(states).toIncludeSameMembers([
+          LogicConditionState.Equal,
+          LogicConditionState.Either,
+        ])
+        expect(states).toBeArrayOfSize(2)
+      })
+    })
+    it('should return valid logic states for binary field types', () => {
+      const states = getApplicableIfStates(BasicField.YesNo)
+      expect(states).toIncludeSameMembers([LogicConditionState.Equal])
+      expect(states).toBeArrayOfSize(1)
+    })
+    it('should return valid logic states for numerical field types', () => {
+      const numericalFields = [
+        BasicField.Number,
+        BasicField.Decimal,
+        BasicField.Rating,
+      ]
+      numericalFields.forEach((fieldType) => {
+        const states = getApplicableIfStates(fieldType)
+        expect(states).toIncludeSameMembers([
+          LogicConditionState.Equal,
+          LogicConditionState.Lte,
+          LogicConditionState.Gte,
+        ])
+        expect(states).toBeArrayOfSize(3)
+      })
+    })
+    it('should return empty array for invalid conditional fields', () => {
+      INVALID_IF_CONDITION_FIELDS.forEach((fieldType) => {
+        const states = getApplicableIfStates(fieldType)
+        expect(states).toStrictEqual([])
+      })
     })
   })
 })
