+++ src/shared/util/logic.ts
@@ -3,6 +3,7 @@ import {
   FieldResponse,
   IClientFieldSchema,
   IConditionSchema,
+  IField,
   IFormDocument,
   ILogicSchema,
   IPreventSubmitLogicSchema,
@@ -48,6 +49,20 @@ const LOGIC_CONDITIONS: LogicCondition[] = [
 export const LOGIC_MAP = new Map<BasicField, LogicConditionState[]>(
   LOGIC_CONDITIONS,
 )
+
+/**
+ * Given a list of form fields, returns only the fields that are
+ * allowed to be present in the if-condition dropdown in the Logic tab.
+ */
+export const getApplicableIfFields = (formFields: IField[]): IField[] =>
+  formFields.filter((field) => !!LOGIC_MAP.get(field.fieldType))
+
+/**
+ * Given a single form field type, returns the applicable logic states for that field type.
+ */
+export const getApplicableIfStates = (
+  fieldType: BasicField,
+): LogicConditionState[] => LOGIC_MAP.get(fieldType) ?? []
 
 type GroupedLogic = Record<string, IConditionSchema[][]>
 export type FieldIdSet = Set<IClientFieldSchema['_id']>
